requirements.txt
Pillow==10.0.0
numpy==1.24.3
wave==0.0.2
pydub==0.25.1
